requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
mysql-connector-python>=8.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0

